Address comments and cleanup
